Fix warning when compiling with i686-w64-mingw32-gcc

+++ src/d_edge.c
@@ -269,7 +269,7 @@ void D_DrawSurfacesPass3()
 	for (surf_t *s = &surfaces[1]; s < surface_p; s++) {
 		if (!s->spans) continue;
 		msurface_t *pface = s->data;
-		u64 is_ent = (u64)(unsigned long long)s->entity & 0xffff000; // FIXME
+		u64 is_ent = (u64)s->entity & 0xffff000; // FIXME
 		if(!(s->flags&SURF_DRAWTURB) && !is_ent) continue;
 		if (pface == 0) continue;
 		d_zistepu = s->d_zistepu;
